+++ orm/src/main/java/org/hibernate/tool/ant/Hbm2JavaExporterTask.java
@@ -7,7 +7,7 @@ package org.hibernate.tool.ant;
 import org.hibernate.tool.api.export.Exporter;
 import org.hibernate.tool.api.export.ExporterFactory;
 import org.hibernate.tool.api.export.ExporterType;
-import org.hibernate.tool.internal.export.pojo.POJOExporter;
+import org.hibernate.tool.api.export.PojoExporter;
 
 /**
  * @author max
