+++ orm/src/test/java/org/hibernate/tool/api/export/ExporterTypeTest.java
@@ -9,7 +9,10 @@ public class ExporterTypeTest {
 	@Test
 	public void testExporterType() {
 		Assert.assertEquals(
-				"org.hibernate.tool.internal.export.pojo.POJOExporter",
+				"org.hibernate.tool.internal.export.common.GenericExporter",
+				ExporterType.GENERIC.className());
+		Assert.assertEquals(
+				"org.hibernate.tool.api.export.PojoExporter",
 				ExporterType.POJO.className());
 	}
